+++ src/cf-serverd.c
@@ -2999,8 +2999,9 @@ static int StatFile(ServerConnectionState *conn, char *sendbuffer, char *ofilena
             cfst.cf_type, (uintmax_t)cfst.cf_mode, (uintmax_t)cfst.cf_lmode, (intmax_t)cfst.cf_uid, (intmax_t)cfst.cf_gid, (long) cfst.cf_size,
             (intmax_t) cfst.cf_atime, (intmax_t) cfst.cf_mtime);
 
-    snprintf(sendbuffer, CF_BUFSIZE, "OK: %d %d %d %d %d %ld %jd %jd %jd %d %d %d %jd",
-             cfst.cf_type, cfst.cf_mode, cfst.cf_lmode, cfst.cf_uid, cfst.cf_gid, (long) cfst.cf_size,
+    snprintf(sendbuffer, CF_BUFSIZE, "OK: %d %ju %ju %ju %ju %ld %jd %jd %jd %d %d %d %jd",
+             cfst.cf_type, (uintmax_t)cfst.cf_mode, (uintmax_t)cfst.cf_lmode,
+             (uintmax_t)cfst.cf_uid, (uintmax_t)cfst.cf_gid, (long) cfst.cf_size,
              (intmax_t) cfst.cf_atime, (intmax_t) cfst.cf_mtime, (intmax_t) cfst.cf_ctime,
              cfst.cf_makeholes, cfst.cf_ino, cfst.cf_nlink, (intmax_t) cfst.cf_dev);
